implemented session silent renewal
